TT#30565 Check for already running instances and abort if so
In end of october we had many running instances because of aborted
ec2-ami-ce Jenkins jobs. To ensure this doesn't happen again
let's check for already running instances if we find any of them
abort within ec2-create-ce before creating any ones.

While this limits us to not have several of them running at the
same time (within the same region and using the same tag!), this
shouldn't be a problem for us for the time being.

Change-Id: I5707f611c6c8a08664a1dbbef08332c36dd2150f