Update patch set 1
Patch Set 1:

> Mika, this is actually strange, as on our side normal reboot worked
 > fine. I have jenkins tests where we install NGCP into chroot AND
 > test where we run ngcp-installer and "reboot" on plain debian VM,
 > here the logs:
 > 
 > https://jenkins.mgm.sipwise.com/job/netinstall-test-ngcp-installer/64/console
 > 
 > I am fine to submit it, so far I would hear Guillem here.
 > 
 > P.S. I suspect Amazon EX2 Debian stretch VM uses default debian
 > stretch systemd version. After the package installation new Sipwise
 > systemd is installed while not active. Those new systemd
 > 237-3~bpo9+1sipwise3 contains guillems fix for "reboot" while I
 > cannot explain how my jenkins test is able to reboot the node. :-(

This feels wrong. If the initial image was using sysvinit. then a similar problem we had at the beginning of the transition is that we were removing too many sysvinit packages, and the compat code in systemd was unable to be used as sysvinit was failing. Could this be the problem here too?

Patch-set: 1