Update patch set 1
Patch Set 1:

> But anyway, the resulting system (ngcp-ce AMI) doesn't correspond
 > to with what we use on our own, so either we ship dbus and/or
 > libpam-systemd everywhere or consider conflicting with it to have
 > consist behavior accross our platforms,  I'm pretty sure other
 > folks who use a default debian-installation to install ngcp-ce also
 > have dbus/libpam-systemd installed (since it's recommended).

I have a general dislike/mistrust of tools like dbus for various possibly-irrational reasons, but I think that it's a kind of lost battle, or one that it's probably not worth fighting for.

So I don't have any strong opinion for or against using dbus/libpam-systemd, but it looks to me as if it would save us some trouble or time.  I am not even sure what's the concrete reason why we're not using it.

So my first option to try to solve this difference would be to have them in all platforms, not spend time to also remove it from cloud images.

In the meantime, this change looks also a double assurance that it will be terminated, so I think that it's a good idea to add it in any case.

Patch-set: 1