Revert "TT#37113 Fix column to check status of instance"
If ip is assigned later ip field in ec2-describe-instances output
is not stable and sometimes it's shown as automatically assigned
one (for instance ip-10-101-136-154.eu-west-1.compute.internal)
(status is in field 6 then) and sometimes it's just empty
(status is in field 5 then). So it looks that having ip assigned
on earlier stage(like it was) makes sense and does
ec2-describe-instances output stable under all circumstances.
So reverting the change here.

This reverts commit 97b7d218e7454c38b3802f6acccaf6ae51e03308.

Change-Id: I26bc165ddfe7c93d9bbaee0ad80811f77af89d6a