Update patch set 1
Patch Set 1:

Nice catch + good fix, thx.

JFTR: The problem is that the *stopped* - and not yet terminated instance (because of we're keeping failing builds) - still holds the elastic IP assignment. (I'm not sure why this never hit us in the past, might be related to the VPC setup or multiple running instances or so.)

Patch-set: 1