TT#47856 ngcp-initial-configuration: use DHCP by default
Instead of hardcoding the IP address for the network interface
that was present at the time of invocation of ngcp-initial-configuration
let's use DHCP instead. This should prevent the system from assigning
and using an IP address that is no longer present, once it was changed
from within the AWS EC2 instance.

Change-Id: I298a2ac1a08f57e45de51c46c02d8832d2db740a