Update patch set 4
Patch Set 4:

> Agreed, but don't we have the same problem by-design also in non-EC2 environments, like people booting outdated systems, installing NGCP CE on it and then only after reboot the kernel matches the system?

Yes, this is a problem but it can't be solved another way but reboot. So we can add such a check but if someone installs ngcp on the system with old kernel he will face the described error.

> (And by *always* rebooting means we make our pipeline intentionally slower)

Yes but the instance is already always rebooted after the configuration so this patchset changes nothing from this POV.
 
> BTW, did you test the code from this review? (Just checking/clarifying)

Sure. Otherwise mr8.2.1 and mr7.5.3 could not be built.

Patch-set: 4
Reviewer: Gerrit User 1000045 <1000045@85315e57-5f03-49c3-83f8-201ae2313a75>