TT#76661 ssh_wrapper: explicitly disable ControlMaster when SSH-ing
We've failing builds for mr6.5.8 and it is yet unclear why we're running
into the timeout of the "Waiting when system is configured" check, because
the system gets configured fine:

| admin@ip-10-0-0-15:~$ cat /etc/sipwise_ngcp_version
| System installed. NGCP version mr6.5.8 on 2020-03-17 09:18:36
| System configured. NGCP version mr6.5.8 on 2020-03-17 09:26:18

We don't see any SSH connection attempts between invocation of
ngcp-initial-configuration and running into the check timeout, when we then
grab /var/log/ngcp-installer*log.

To avoid having an SSH connection running with ControlMaster let's
see whether explicitly disabling this modifies the behavior.

Change-Id: I7af5e7113b5c031acc5cdf78c9904dabf00e9ada