TT#76661 Fix usage of ngcp version detection
The variable "${vmversion}" is not available within this script and comes
from a c/p bug introduced in commit 8c26300a.

This fixes a build failure for all NGCP releases that don't provide status
information on port 4242 (which is available only with >=mr7.5) but instead
need to be checked via /etc/sipwise_ngcp_version.

Change-Id: Ic2b233a69737c5424e30d6971ed642c19b3adac7